Moved routing up to global and refactored pre-checkin to use.
