+++ src/applications/check-in/hooks/tests/useFormRouting/TestComponent.jsx
@@ -2,6 +2,17 @@ import React from 'react';
 import { useFormRouting } from '../../useFormRouting';
 
 export default function TestComponent({ router }) {
+  const URLS = Object.freeze({
+    CONFIRMATION: 'complete',
+    DEMOGRAPHICS: 'contact-information',
+    EMERGENCY_CONTACT: 'emergency-contact',
+    ERROR: 'error',
+    INTRODUCTION: 'introduction',
+    LANDING: '',
+    NEXT_OF_KIN: 'next-of-kin',
+    SEE_STAFF: 'see-staff',
+    VERIFY: 'verify',
+  });
   const {
     currentPage,
     goToPreviousPage,
@@ -9,7 +20,7 @@ export default function TestComponent({ router }) {
     goToErrorPage,
     jumpToPage,
     pages,
-  } = useFormRouting(router);
+  } = useFormRouting(router, URLS);
   return (
     <div>
       <h1>Test component for the useFormRouting hook</h1>
