+++ src/applications/check-in/pre-check-in/actions/index.js
@@ -1,15 +1,3 @@
-export const INIT_FORM = 'INIT_FORM';
-
-export const createInitFormAction = ({ pages, firstPage }) => {
-  return {
-    type: INIT_FORM,
-    payload: {
-      pages,
-      currentPage: firstPage,
-    },
-  };
-};
-
 export const SET_SESSION = 'SET_SESSION';
 
 export const createSetSession = ({ token, permissions }) => {
@@ -19,15 +7,6 @@ export const createSetSession = ({ token, permissions }) => {
       token,
       permissions,
     },
-  };
-};
-
-export const GO_TO_NEXT_PAGE = 'GO_TO_NEXT_PAGE';
-
-export const createGoToNextPageAction = ({ nextPage }) => {
-  return {
-    type: GO_TO_NEXT_PAGE,
-    payload: { nextPage },
   };
 };
