+++ src/applications/check-in/pre-check-in/containers/withAuthorization.jsx
@@ -1,9 +1,9 @@
 import React, { useMemo, useEffect } from 'react';
 import { useSelector } from 'react-redux';
 
-import { makeSelectCurrentContext } from '../selectors';
+import { makeSelectCurrentContext } from '../../selectors';
 
-import { useFormRouting } from '../hooks/useFormRouting';
+import { useFormRouting } from '../../hooks/useFormRouting';
 import { useSessionStorage } from '../hooks/useSessionStorage';
 
 import { SCOPES } from '../../utils/token-format-validator';
@@ -15,7 +15,7 @@ const withAuthorization = Component => {
     const selectCurrentContext = useMemo(makeSelectCurrentContext, []);
     const { token, permissions } = useSelector(selectCurrentContext);
 
-    const { jumpToPage, goToErrorPage } = useFormRouting(router);
+    const { jumpToPage, goToErrorPage } = useFormRouting(router, URLS);
     const { getCurrentToken } = useSessionStorage();
 
     useEffect(
