+++ src/applications/check-in/pre-check-in/pages/Demographics/index.jsx
@@ -2,21 +2,23 @@ import React, { useCallback, useEffect, useMemo } from 'react';
 import { focusElement } from 'platform/utilities/ui';
 import { useDispatch, useSelector } from 'react-redux';
 import BackToHome from '../../components/BackToHome';
-import { useFormRouting } from '../../hooks/useFormRouting';
+import { useFormRouting } from '../../../hooks/useFormRouting';
 import PropTypes from 'prop-types';
 import Footer from '../../components/Footer';
 import BackButton from '../../components/BackButton';
 import DemographicsDisplay from '../../../components/pages/demographics/DemographicsDisplay';
 import recordEvent from 'platform/monitoring/record-event';
 import { recordAnswer } from '../../actions';
+import { URLS } from '../../utils/navigation';
 
-import { makeSelectVeteranData } from '../../selectors';
+import { makeSelectVeteranData } from '../../../selectors';
 
 const Demographics = props => {
   const dispatch = useDispatch();
   const { router } = props;
   const { goToNextPage, goToPreviousPage, currentPage } = useFormRouting(
     router,
+    URLS,
   );
   useEffect(() => {
     focusElement('h1');
