+++ src/applications/check-in/pre-check-in/pages/NextOfKin/index.jsx
@@ -14,13 +14,14 @@ import BackToHome from '../../components/BackToHome';
 import Footer from '../../components/Footer';
 import NextOfKinDisplay from '../../../components/pages/nextOfKin/NextOfKinDisplay';
 
-import { useFormRouting } from '../../hooks/useFormRouting';
+import { useFormRouting } from '../../../hooks/useFormRouting';
+import { URLS } from '../../utils/navigation';
 
 import {
   makeSelectCurrentContext,
   makeSelectVeteranData,
   makeSelectForm,
-} from '../../selectors';
+} from '../../../selectors';
 
 import { makeSelectFeatureToggles } from '../../../utils/selectors/feature-toggles';
 
@@ -50,7 +51,7 @@ const NextOfKin = props => {
     goToErrorPage,
     goToNextPage,
     goToPreviousPage,
-  } = useFormRouting(router);
+  } = useFormRouting(router, URLS);
 
   useEffect(() => {
     focusElement('h1');
