+++ src/applications/check-in/pre-check-in/reducers/index.js
@@ -11,13 +11,9 @@ const initialState = {
   },
 };
 
-import {
-  GO_TO_NEXT_PAGE,
-  INIT_FORM,
-  RECORD_ANSWER,
-  SET_SESSION,
-  SET_VETERAN_DATA,
-} from '../actions';
+import { RECORD_ANSWER, SET_SESSION, SET_VETERAN_DATA } from '../actions';
+
+import { GO_TO_NEXT_PAGE, INIT_FORM } from '../../actions';
 
 const preCheckInReducer = (state = initialState, action) => {
   switch (action.type) {
@@ -64,5 +60,5 @@ const preCheckInReducer = (state = initialState, action) => {
 };
 
 export default {
-  preCheckInData: preCheckInReducer,
+  checkInData: preCheckInReducer,
 };
